Login Improved and password hash added

+++ admin/includes/view_all_posts.php
@@ -124,7 +124,7 @@
                                     <a class='btn btn-primary' href='posts.php?source=edit_post&post_id={$post_id}'>Edit</a>
                                 </td>";
                             echo "<td align='center'>
-                                    <a class='btn btn-danger' href='posts.php?delete={$post_id}'>Delete</a>
+                                    <a class='btn btn-danger' onClick=\"javascript: return confirm('Are you sure you want to delete this post'); \" href='posts.php?delete={$post_id}'>Delete</a>
                                 </td>";
                         echo "</tr>";
                     }
